refact: reafatora profundamente o projeto.
- Reve usecases
- Refatora Eventos
- Cria abstrações
- Prepara para concorrencia
- Remove entidades UploadSession e Chunk

+++ domain/src/main/java/com/arqivame/storage/domain/file/service/ChunkCalculatorService.java
@@ -19,14 +19,14 @@ public final class ChunkCalculatorService {
         final long fullChunks = fileSize / maxAllowedChunkSize;
         final long hasPartialChunk = fileSize % maxAllowedChunkSize != 0 ? 1 : 0;
 
-        final long totalChunks = fullChunks + hasPartialChunk;
+        final int totalChunks = (int) (fullChunks + hasPartialChunk);
         final long lastChunkSize = hasPartialChunk == 1 ? fileSize % maxAllowedChunkSize : maxAllowedChunkSize;
 
         return new ChunkCalculationResult(totalChunks, maxAllowedChunkSize, lastChunkSize);
 
     }
 
-    public record ChunkCalculationResult(Long totalChunks, Long chunkSize, Long lastChunkSize) {
+    public record ChunkCalculationResult(Integer totalChunks, Long chunkSize, Long lastChunkSize) {
     }
 
 }
